Make a `main` util
