+++ tools/floretta/Dockerfile
@@ -18,7 +18,12 @@ RUN chmod +x bin/floretta
 
 # `COPY` the rest of the files.
 COPY tools/floretta tools/floretta
+COPY package.json .
+COPY js/common js/common
 COPY js/floretta js/floretta
+
+# Create `node_modules`.
+RUN npm install --omit dev
 
 ENTRYPOINT ["node", "--disable-warning=ExperimentalWarning", "js/floretta/run.ts"]
 LABEL org.opencontainers.image.source=https://github.com/gradbench/gradbench
